Clean up lib

+++ openCVLibrary300/src/main/java/org/opencv/core/TermCriteria.java
@@ -12,7 +12,8 @@ public class TermCriteria {
      */
     public static final int MAX_ITER = COUNT;
     /**
-     * The desired accuracy threshold or change in parameters at which the iterative algorithm is terminated.
+     * The desired accuracy threshold or change in parameters at which the iterative algorithm is
+     * terminated.
      */
     public static final int EPS = 2;
 
@@ -23,12 +24,9 @@ public class TermCriteria {
     /**
      * Termination criteria for iterative algorithms.
      *
-     * @param type
-     *            the type of termination criteria: COUNT, EPS or COUNT + EPS.
-     * @param maxCount
-     *            the maximum number of iterations/elements.
-     * @param epsilon
-     *            the desired accuracy.
+     * @param type the type of termination criteria: COUNT, EPS or COUNT + EPS.
+     * @param maxCount the maximum number of iterations/elements.
+     * @param epsilon the desired accuracy.
      */
     public TermCriteria(int type, int maxCount, double epsilon) {
         this.type = type;
@@ -51,7 +49,7 @@ public class TermCriteria {
         if (vals != null) {
             type = vals.length > 0 ? (int) vals[0] : 0;
             maxCount = vals.length > 1 ? (int) vals[1] : 0;
-            epsilon = vals.length > 2 ? (double) vals[2] : 0;
+            epsilon = vals.length > 2 ? vals[2] : 0;
         } else {
             type = 0;
             maxCount = 0;
